+++ lib/bwu_datastore_launcher.dart
@@ -8,5 +8,3 @@ import 'package:bwu_utils_server/network/network.dart';
 part 'src/server.dart';
 part 'src/datastore_local_dev_server.dart';
 part 'src/appengine_api_server.dart';
-//part 'src/util.dart';
-