+++ test/launch_datastore_local_dev_server_test.dart
@@ -1,20 +1,14 @@
 library bwu_datastore_launcher.test.launch_datastore_local_dev_server;
 
-import 'dart:io' as io;
-import 'dart:async' show Future;
 import 'package:path/path.dart' as path;
 import 'package:unittest/unittest.dart';
-import 'package:unittest/vm_config.dart';
 import 'package:bwu_datastore_launcher/bwu_datastore_launcher.dart';
 import 'package:bwu_utils_server/package/package.dart';
 
-main() async {
-  useVMConfiguration();
-
-  group('launch local dev server', () {
-
+main() {
+  group('launch Datastore Local Dev Server', () {
     test('start and remoteSuthdown', () {
-      var exitCalled = expectAsync((){});
+      var exitCalled = expectAsync(() {});
 
       // Create an instance of the server launcher.
       final server = new DatastoreLocalDevServer('connect',
@@ -32,7 +26,7 @@ main() async {
         expect(success, isTrue);
 
         // when done launch the Gcloud Datastore Local Development Server
-        server.onExit.first.then((code) {
+        return server.onExit.first.then((code) {
           return server
               .start(allowRemoteShutdown: true, doStoreOnDisk: false)
               .then((success) {
@@ -43,15 +37,54 @@ main() async {
               exitCalled();
             });
 
-            // We need to wait a little until the server is able to receive
-            // requests.
-            return new Future.delayed(new Duration(seconds: 3),
-                // () => server.kill(io.ProcessSignal.SIGTERM))
-                // Darts Process kill doesn't kill child processes, therefore we
-                // use the `remoteShutdown` feature of the server to not keep
-                // unnecessary server processes running.
-                () => server.remoteShutdown()).then(
-                (success) => expect(success, isTrue));
+            return server
+                .remoteShutdown()
+                .then((success) => expect(success, isTrue));
+          });
+        });
+      });
+    });
+
+    test('start and remoteSuthdown without delay should fail', () {
+      var exitCalled = expectAsync(() {});
+
+      // Create an instance of the server launcher.
+      final server = new DatastoreLocalDevServer('connect',
+          workingDirectory: path.join(
+              packageRoot().absolute.path, 'test/tmp_data'),
+          // `gcd` uses the `JAVA` environment variable to find the Java
+          // executable. We make it to point to Java 7 because `gcd` has issues
+          // with Java 8.
+          environment: <String, String>{
+        'JAVA': '/usr/lib/jvm/java-7-openjdk-amd64/bin/java'
+      },
+          minUpTimeBeforeShutdown: new Duration(seconds: 0));
+
+      // create the datastore directory
+      return server.create('test', deleteExisting: true).then((success) {
+        expect(success, isTrue);
+
+        // when done launch the Gcloud Datastore Local Development Server
+        return server.onExit.first.then((code) {
+          return server
+              .start(allowRemoteShutdown: true, doStoreOnDisk: false)
+              .then((success) {
+            expect(success, isTrue);
+
+            server.onExit.first.then((code) {
+              expect(code, equals(0));
+              exitCalled();
+            });
+
+            return server
+                .remoteShutdown()
+                .then((success) => expect(success, isFalse))
+                .then((_) {
+              server.minUpTimeBeforeShutdown = new Duration(seconds: 3);
+              return server
+                  .remoteShutdown()
+                  .then((success) => expect(success, isTrue));
+            });
           });
         });
       });